requirements.txt
opencv-python
numpy
stegano
Pillow